+++ opentelemetry-extra/src/OpenTelemetry/FileExporter.hs
@@ -4,9 +4,11 @@ module OpenTelemetry.FileExporter where
 
 import Data.Function
 import qualified Data.HashMap.Strict as HM
-import Data.List
+import Data.List (intersperse)
 import qualified Data.Text as T
 import OpenTelemetry.Common
+import OpenTelemetry.Exporter
+import OpenTelemetry.SpanContext
 import System.IO
 import Text.Printf
 import Text.Read