+++ opentelemetry-wai/src/OpenTelemetry/Network/Wai/Middleware.hs
@@ -2,11 +2,10 @@
 
 module OpenTelemetry.Network.Wai.Middleware where
 
-import qualified Data.Text as T
-import qualified Data.Text.Encoding as T
+import qualified Data.ByteString.Char8 as BS8
 import Network.HTTP.Types
 import Network.Wai
-import OpenTelemetry.Implicit
+import OpenTelemetry.Eventlog
 import OpenTelemetry.Propagation
 
 -- Semantic conventions for HTTP spans:
@@ -18,11 +17,11 @@ middleware app = \req sendResp -> do
     case propagateFromHeaders w3cTraceContext (requestHeaders req) of
       Just ctx -> setParentSpanContext ctx
       _ -> pure ()
-    setTag "span.kind" ("server" :: T.Text)
-    setTag "component" ("http" :: T.Text)
-    setTag "http.method" $ T.decodeUtf8 (requestMethod req)
-    setTag "http.target" $ T.decodeUtf8 (rawPathInfo req)
-    setTag "http.flavor" $ show (httpVersion req)
+    setTag "span.kind" "server"
+    setTag "component" "http"
+    setTag "http.method" $ requestMethod req
+    setTag "http.target" $ rawPathInfo req
+    setTag "http.flavor" $ BS8.pack $ show (httpVersion req)
     app req $ \resp -> do
-      setTag "http.status_code" (statusCode $ responseStatus resp)
+      setTag "http.status_code" (BS8.pack $ show $ statusCode $ responseStatus resp)
       sendResp resp